refactor Phoebe to use ILoggerClient
