+++ Phoebe/Phoebe.Android.csproj
@@ -62,6 +62,7 @@
     <Compile Include="Data\ViewModels\NewProjectViewModel.cs" />
     <Compile Include="Data\Utils\TimeEntryHolder.cs" />
     <Compile Include="Data\Views\TimeEntriesCollectionView.cs" />
+    <Compile Include="Logging\ILoggerClient.cs" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Contrib\" />
